requirements.txt
asgiref==3.5.2
Django==4.0.5
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
environs==9.5.0
marshmallow==3.16.0
packaging==21.3
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
